backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.25.0
fastf1>=3.0.0
